+++ tests/testthat/test-read-survey.R
@@ -1,7 +1,7 @@
-test_that("readSurvey() reads data in qualtrics standard and legacy format and converts columns", { # nolint
+test_that("read_survey() reads data in qualtrics standard and legacy format and converts columns", { # nolint
 
   # Test if can read standard format
-  survey <- suppressWarnings(qualtRics::readSurvey("files/sample.csv"))
+  survey <- suppressWarnings(qualtRics::read_survey("files/sample.csv"))
   # Tests
   expect_equal(dim(survey)[1], 1)
   expect_equal(dim(survey)[2], 20)
@@ -9,8 +9,8 @@ test_that("readSurvey() reads data in qualtrics standard and legacy format and c
   expect_true(is.numeric(survey$LocationLatitude))
 
   # Test if can read legacy format
-  survey_legacy <- suppressWarnings(qualtRics::readSurvey("files/sample_legacy.csv", # nolint
-    legacyFormat = TRUE
+  survey_legacy <- suppressWarnings(qualtRics::read_survey("files/sample_legacy.csv", # nolint
+    legacy = TRUE
   ))
   # Tests
   expect_equal(dim(survey_legacy)[1], 1)
@@ -21,7 +21,7 @@ test_that("readSurvey() reads data in qualtrics standard and legacy format and c
 
 test_that("Survey exists to read from disk", {
   expect_error(
-    qualtRics::readSurvey("/users/julia/desktop/error.csv"),
+    qualtRics::read_survey("/users/julia/desktop/error.csv"),
     "does not exist"
   )
 })